优化所有实体类主键，禁止了主键自增

+++ IT-pojo/src/main/java/com/xiaoyan/vo/AiDialogVO.java
@@ -17,9 +17,9 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class AiDialogVO implements Serializable {
 
-    private Integer id;
+    private Long id;
 
-    private Integer sessionId;
+    private Long sessionId;
 
     private String  senderType;
